chore: extract `BookSelection`
